Restructure segment rendering methods
Theme.get_segments() is now a generator which returns segment copies,
and it no longer handles segment content replacement.

Renderer.render() accepts a theme and a segments parameter in order to
override the default actions (calling self.get_theme() and
self.get_segments()). This is necessary to be able to cache themes and
segment contents somewhere else and provide the cached data to the
render method. The render method now also handles removing excluded
segments, which was previously handled in Theme.get_segments().

Finally, VimRenderer.render() caches all theme and segment data and
provides this data to the renderer for non-current windows.

Closes #12.