Added virtcol_current segment. Also made it default
I doubt anybody here ever wanted to see byte offset. When it comes to 
a questions like alignment, forced text width and so on only virtual columns 
matter. It would be good to optionally take concealed characters into account, 
but vim does not provide a way to do so (well, except recent screencol() added 
solely for testing which would be hard, but not impossible, to adapt to such 
needs).