Add file_vcs_status segment, using powerline.lib.vcs instead of fugitive
Using pure-python implementation allows to use the same code in the
terminal segments (not implemented currently). It also should be more
effective. Git module is able to use pygit2 instead of command-line git
which is much faster. Mercurial module assumes that mercurial is always
accessible as a python module.

Code for repository (not file) status is untested. It exists solely for
terminal segments.

Closes #26.