Highlight VCS flags differently
"M" flag is highlighted in yellow, "A" flag is highlighted in green,
other flags are highlighted in red. This also corresponds with the
highlighting in the output of `git status`.

Whitespace is also trimmed from the flags as the columns don't matter
much in vim statuslines, and empty columns only introduce whitespace
that looks like a bug.