Move terminal and tmux segments into common module
This should ideally be renamed to something else since it's not strictly
an extension. A better module naming scheme could be something like:

	powerline.segments.ext.{extension}
	powerline.segments.common

This would move all segments out of their respective extension
directory, and might be a bit confusing. The ext directories may also
need some changes to make this work properly.