Add zsh prompt script for sourcing in zshrc
This script sets a couple of environment variables in tmux if it's
running, and provides the last exit code and last pipe status to
powerline-prompt. It also traps SIGWINCH and sets the window width as an
env variable in tmux. The default prompt has been split in two and the
git branch is moved to RPS1.

The script must be sourced with an absolute path for now:

    . /path/to/powerline/bindings/zsh/prompt.zsh

Statusline cropping has not been implemented in powerline-tmux yet.

Closes #82.
Closes #83.
Refs #90.