Restored export as it makes sense in bash
Bash has weird initialization: bashrc is sourced in non-login interactive shell, 
but not in login interactive shell; profile is sourced in login interactive 
shell, but not in non-login interactive, thus exporting in profile does make 
sense. In zsh case zshrc is always sourced when shell is interactive and 
exporting thus makes no sense.