Made zsh bindings restore prompt
Prompt is restored when zsh/zpython module is unloaded: as ZPYTHON_* variable is 
readonly (unless __call__ method is defined) it is currently impossible to 
delete it without unloading module, though prompt can be restored independent of 
variable deletion. This zsh/zpython behavior is currently considered to be 
a feature. If __call__ method is defined prompt will also be restored when 
variable is unset