Optimize cterm_to_hex
Dictionary is an overkill here. Tuple is faster. Also now the whole tuple can be
seen on the 1/4 of my screen (32x119 characters) and colors are easier to find
(useful only when seeking for normal color names).

Again, it would be goot to ignore flake8 in this case. Without comments it
complains about it is not easy to find requested color by hand