Fix vcs file status segment
The file status cache was being invalidated for only a single file in
the repo when the dirstate file changed, instead of all files. This
meant that having multiple modified open files in vim and then
committing would cause the status to be updated for only one of the
files.