Move some functions from VimL to python code
Should be faster with new interfaces, but that is not the case

Should remove zero that can be viewed temporary in place of statusline in new 
windows

Target: ressurect setup function like source_plugin (but in place of sourcing 
something code is moved to python) to make it possible to do the following 
sequence:

1. Install powerline via pip
2. Add a line to the vimrc like

        py from powerline.vim import setup as setup_powerline; setup_powerline(); del setup_powerline

3. See new fancy statusline

Currently there are no convenient options to use powerline installed by pip in 
vim. powerline/bindings/vim/plugin/powerline.vim will do what it does: check for 
appropriate python versions available, fix sys.path and so on, but it won’t 
create autocommands, set options or define PowerlinePyeval function.

Note: may be rebased on top of the develop.