Clean up package function for python3
``|| return 1`` is very old syntax and is no longer needed. And, the ``install`` command can handle directory creation and file installation in the same command which cuts a great deal of unnecessary commands out.

Also, remember, ``msg2`` in the package function will be printed during the making of the package, not during the installation. Thus, instead of ``msg2`` (as they really serve only as documentation), comments should be preferred.