Dont log inotify missing error when using branch coloring
The error message about inotify being missing when branch coloring is
enabled was being logged for every directory and for every invocation
of powerline in a shell, without powerline-daemon. Dont log it, since
the log is printed to stderr when using powerline in a shell without
powerline-daemon. Also improve the error message on OS X. Fixes #578