Update powerline-release script to use live ebuild as the base
Reasoning: I am going to add man pages support to the live ebuild, but it does 
not make sense to add this to the 1.0 ebuild previously used as the base. Since 
new features will likely continue to appear in the live ebuild powerline-release 
script was modified to do the right thing under new circumstances.