Do not bother with testing or setting POWERLINE_CONFIG_COMMAND
Instead just test whether using $POWERLINE_CONFIG_COMMAND succeeds and if it 
does not run just `powerline-config tmux setup`. Problem is that tests in vterm 
fail *not* because sleeping time is too short, but because `powerline-config 
tmux setup` for some reason is not launched at all. I am not sure whether this 
construct is going to fix things.

Also ref #875. I am very unsure this commit will fix the issue, but since I do 
not know what is the reason for this issue at all… who knows.