Add git merge/rebase to git branch segment.
Prior to using powerline with bash, I used $(__git_ps1) to add git
branch info to my prompt.  __git_ps1 includes the branch name, but it
also provides some additional status information that is incredibly
helpful for staying oriented when in the middle of some git operations
like a merge or rebase.

This patch implements something similar for the git implementation of
the vcs branch segment.  For example, if you're in the middle of an
interactive rebase on branch "master", the branch segment will now
show "master (REBASE-i)".