Fix #1483 -- [Battery - Linux] Fix AC status fetch
Fix a bug where the `online` property is checked inside battery
instead of AC.

As specified in issue #1483, `online` is found only inside
`/sys/class/power_supply/AC<x>/` directories, and equals `1` when cable is
plugged, and `0` when unplugged.

Edit: updated with --amend according to ZyX-I hints in order to work with
powerline daemon.