provide stash counter
Some version control systems have an area where changes can be stored as
temporary work in progress instead publishing them to the history.

This segment is intended to display their count, reminding the developer
about open ends in the current repository.

The underlying implementation is VCS-specific and has to be provided by
the respective library module.