Added note about laststatus
Users may be confused when installing powerline as it does not appear by default in single-window mode. This change alerts users that they should not expect to see the status line and points them to the help documents for laststatus so that they can activate it in all cases if they want.