Render cpu_load_percent segment even when psutil returns 0.0
Although the docs for psutil say 0.0 is a nonsense value that's returned
first and supposed to be ignored, but on a relatively idle system it
will actually return 0.0 occasionally after that. So at the expense of
displaying that "nonsense value" when the statusbar first loads, this
change gets rid of the visual disruptions of having the segment pop out
and back in periodically.

Fixes #2110.